+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+* Fix EncryptionMode.NONE enum value to match value expected by B2 API
+
 ## [1.24.0] - 2023-08-31
 
 ### Added